requirements.txt
youtube_transcript_api==1.0.3
google.generativeai==0.8.4
beautifulsoup4==4.13.3
requests==2.25.1
python-dotenv==1.1.0
flask_cors==5.0.1
flask==3.1.0